mkvalidator: compare DocType and read DocType the right way
DocType contains the profile that was used for creating the file. Read
DocType is the minimum profile a player must support in order to play
this file. Therefore read DocType can be smaller than DocType which
means that a player can skip unknown elements from higher profiles but
it should still be able to play the file just fine.

Read DocType can logically not be greater than DocType: if the file uses
elements from at most version X then it wouldn't make sense to require a
player that supports a profile greater than X.