libmatroska2: 0 duration frames can be keyframes
The part of the Matroska specifications that claimed that Blocks from a
BlockGroup with duration zero are no keyframes has been dropped in
9313a00f7391a285f95e9d6f0a5f8f3f80ed4779. This commit brings
libmatroska2 in line with this change.

The earlier version e.g. made mkvalidator warn about files muxed by
FFmpeg that contained pgs subtitles (which are muxed as zero-duration
subtitles, which is spec-compliant).

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@googlemail.com>