libebml2: fix potential reading overflows
We should never end up reading an element with an incorrect value. But in case
we do, return an error.

We should not try to allocate more than SIZE_MAX, on 32 bits it will be cropped
to a lower value and end up writing in incorrect data. Technically the file may
be correct, but it can't be read (in memory) on 32 bits system.